+++ stream-chat-android-pushprovider-firebase/src/main/java/io/getstream/chat/android/pushprovider/firebase/ChatFirebaseMessagingService.kt
@@ -3,12 +3,7 @@ package io.getstream.chat.android.pushprovider.firebase
 import android.util.Log
 import com.google.firebase.messaging.FirebaseMessagingService
 import com.google.firebase.messaging.RemoteMessage
-import io.getstream.chat.android.client.ChatClient
 import io.getstream.chat.android.client.logger.ChatLogger
-import io.getstream.chat.android.client.models.Device
-import io.getstream.chat.android.client.models.PushMessage
-import io.getstream.chat.android.client.models.PushProvider
-import kotlin.jvm.Throws
 
 internal class ChatFirebaseMessagingService : FirebaseMessagingService() {
     private val logger = ChatLogger.get("ChatFirebaseMessagingService")
@@ -16,11 +11,9 @@ internal class ChatFirebaseMessagingService : FirebaseMessagingService() {
     override fun onMessageReceived(remoteMessage: RemoteMessage) {
         logger.logD("onMessageReceived(): $remoteMessage")
         try {
-            remoteMessage.takeIf { it.isValid() }
-                ?.toPushMessage()
-                ?.run { ChatClient.handlePushMessage(this) }
+            FirebaseMessagingDelegate.handleRemoteMessage(remoteMessage)
         } catch (exception: IllegalStateException) {
-            Log.e(TAG, "Error while handling remote message: ${exception.message}")
+            Log.e(TAG, "Error while handling remote message", exception)
         } finally {
             stopSelf()
         }
@@ -28,14 +21,9 @@ internal class ChatFirebaseMessagingService : FirebaseMessagingService() {
 
     override fun onNewToken(token: String) {
         try {
-            ChatClient.setDevice(
-                Device(
-                    token = token,
-                    pushProvider = PushProvider.FIREBASE,
-                )
-            )
+            FirebaseMessagingDelegate.registerFirebaseToken(token)
         } catch (exception: IllegalStateException) {
-            Log.e(TAG, "Error while handling remote message: ${exception.message}")
+            Log.e(TAG, "Error while registering Firebase Token", exception)
         }
     }
 
@@ -43,18 +31,3 @@ internal class ChatFirebaseMessagingService : FirebaseMessagingService() {
         private const val TAG = "Chat:"
     }
 }
-
-@Throws(IllegalStateException::class)
-private fun RemoteMessage.toPushMessage() = when (isValid()) {
-    true -> PushMessage(
-        channelId = data["channel_id"]!!,
-        messageId = data["message_id"]!!,
-        channelType = data["channel_type"]!!,
-    )
-    else -> throw IllegalStateException("RemoteMessage doesn't contains needed data")
-}
-
-private fun RemoteMessage.isValid() =
-    !data["channel_id"].isNullOrBlank() &&
-        !data["message_id"].isNullOrBlank() &&
-        !data["channel_type"].isNullOrBlank()